requirements.txt
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
